Fix Series.__getitem__ with a slice
* Combine codepaths for `DataFrame` and `Series`
