+++ modin/pandas/dataframe.py
@@ -12,7 +12,7 @@ from pandas.core.dtypes.common import (
     is_numeric_dtype,
 )
 from pandas.core.index import ensure_index_from_sequences
-from pandas.core.indexing import check_bool_indexer, convert_to_index_sliceable
+from pandas.core.indexing import check_bool_indexer
 from pandas.util._validators import validate_bool_kwarg
 
 import itertools
@@ -1825,11 +1825,6 @@ class DataFrame(BasePandasDataset):
                 return self._getitem_column(key)
         except (KeyError, ValueError, TypeError):
             pass
-        # see if we can slice the rows
-        # This lets us reuse code in Pandas to error check
-        indexer = convert_to_index_sliceable(pandas.DataFrame(index=self.index), key)
-        if indexer is not None:
-            return self._getitem_slice(indexer)
         if isinstance(key, (Series, np.ndarray, pandas.Index, list)):
             return self._getitem_array(key)
         elif isinstance(key, DataFrame):
@@ -1885,28 +1880,6 @@ class DataFrame(BasePandasDataset):
             return DataFrame(
                 query_compiler=self._query_compiler.getitem_column_array(key)
             )
-
-    def _getitem_slice(self, key):
-        # If there is no step, we can fasttrack the codepath to use existing logic from
-        # head and tail, which is already pretty fast.
-        if key.step is None:
-            if key.start is None and key.stop is None:
-                return self.copy()
-
-            def compute_offset(value):
-                return value - len(self) if value >= 0 else value
-
-            # Head is a negative number, Tail is a positive number
-            if key.start is None:
-                return self.head(compute_offset(key.stop))
-            elif key.stop is None:
-                return self.tail(-compute_offset(key.start))
-            return self.head(compute_offset(key.stop)).tail(-compute_offset(key.start))
-        # We convert to a RangeIndex because getitem_row_array is expecting a list
-        # of indices, and RangeIndex will give us the exact indices of each boolean
-        # requested.
-        key = pandas.RangeIndex(len(self.index))[key]
-        return DataFrame(query_compiler=self._query_compiler.getitem_row_array(key))
 
     def __getattr__(self, key):
         """After regular attribute access, looks up the name in the columns
